Refactor `fetchers` core into multiple submodules
