+++ src/core/fetchers/internal/constants.ts
@@ -1,0 +1,3 @@
+export const PER_PAGE_COUNT = 100;
+
+export const CONCURRENCY_LIMIT = 5;
